fix(spinlock.h): correctly use uint8_t
The atomic primitive we're using requires 8-bit targets.

+++ kernel/core/spinlock.h
@@ -9,7 +9,7 @@
 // Contains a lock we will spin on.
 struct spinlock {
 	// The value controlling spin locking.
-	volatile uint64_t value;
+	volatile uint8_t value;
 };
 
 // Use this macro to intialize a static spinlock.
